refactor: rename interface to queue_interface
