+++ app/topic_prompt/contextualize.py
@@ -1,7 +1,7 @@
 """
 Provide context for a source
 """
-from app.sefaria_interface.topic_prompt_source import TopicPromptSource
+from app.queue_interface.topic_prompt_source import TopicPromptSource
 from app.util.general import get_source_text_with_fallback, get_by_xml_tag
 import re
