+++ app/topic_prompt/html_formatter.py
@@ -2,7 +2,7 @@ from typing import List
 from toprompt import TopromptOptions, Toprompt
 from app.util.sefaria_specific import get_raw_ref_text
 from abstract_formatter import AbstractFormatter
-from app.sefaria_interface.topic import Topic
+from app.queue_interface.topic import Topic
 import math
