+++ app/topic_prompt/topic_prompt_generator.py
@@ -3,9 +3,9 @@ from loguru import logger
 
 from tqdm import tqdm
 from typing import List
-from app.sefaria_interface.topic_prompt_input import TopicPromptInput
-from app.sefaria_interface.topic_prompt_source import TopicPromptSource
-from app.sefaria_interface.topic import Topic
+from app.queue_interface.topic_prompt_input import TopicPromptInput
+from app.queue_interface.topic_prompt_source import TopicPromptSource
+from app.queue_interface.topic import Topic
 from app.topic_prompt.toprompt_llm_prompt import TopromptLLMPrompt, get_output_parser
 from app.topic_prompt.toprompt import Toprompt, TopromptOptions
 from app.topic_prompt.differentiate_writing import repeated_phrase
