+++ app/topic_prompt/uniqueness_of_source.py
@@ -7,8 +7,8 @@ from functools import reduce
 from typing import List
 from app.util.general import get_source_text_with_fallback
 
-from app.sefaria_interface.topic_prompt_source import TopicPromptSource
-from app.sefaria_interface.topic import Topic
+from app.queue_interface.topic_prompt_source import TopicPromptSource
+from app.queue_interface.topic import Topic
 
 from langchain.prompts import PromptTemplate
 from app.basic_langchain.schema import HumanMessage, SystemMessage
