+++ app/util/general.py
@@ -1,4 +1,4 @@
-from app.sefaria_interface.topic_prompt_source import TopicPromptSource
+from app.queue_interface.topic_prompt_source import TopicPromptSource
 import diff_match_patch
 import re
